% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scBinning.R
\name{sc.binning}
\alias{sc.binning}
\title{Numeric binning for Credit Scoring}
\usage{
sc.binning(data, target, n = 10, p = 3, thres = 0.5,
  freqCut = 95/5, uniqueCut = 10, best = TRUE, parallel = FALSE)
}
\arguments{
\item{data}{A data frame which contains target varible as well as predictor variables.}

\item{target}{Target variable name.}

\item{n}{Number of bootstrap iterations. Default 10 times.}

\item{p}{The minimum percentage of observation per bin. Default 3\%.}

\item{thres}{Threshold differences of target between bins. Default 0.5\%.}

\item{freqCut}{Utilizing \code{\link[caret:nearZeroVar]{nearZeroVar}} function. The cutoff for the ratio of the most common value to the second most common value. Default 95/5.}

\item{uniqueCut}{Utilizing \code{\link[caret:nearZeroVar]{nearZeroVar}} function. The cutoff for the percentage of distinct values out of the number of total samples. Default 10\%.}

\item{best}{A logical scalar. Use different methods which maximize IV. Default TRUE.}

\item{parallel}{A logical scalar. Use parallel backend. Default FALSE.}
}
\value{
The output is a list of cut plan which can be applied to the orginal data frame via
  the \code{\link{predict}} function.
  The user can also update the cut plan via the \code{\link{update}} function.
}
\description{
This is an easy numerical binning solution for credit scorecard build. It is designed to
choose the optimal binning solution by utilizing the \href{https://cran.r-project.org/package=partykit}{Recursive Partitioning}.
This will only bin numeric or integer variables and ignore factor or character variables.
}
\examples{
\dontrun{
# Load library
library(easysc)

# Generate a cut plan which maximize IV via 500 bootstrap resampling
cut.plan <- sc.binning(data = df, target = BAD, n = 500, p = 5, best = TRUE, parallel = TRUE)
# Update the cut plan
update(cut.plan, AGE = c(20, 30, 40))
# Apply to the data frame
predict(cut.plan, df, keepTarget = TRUE)
}
}
