% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scCatBinning.R
\name{sc.cat.binning}
\alias{sc.cat.binning}
\title{Categorical binning for Credit Scoring}
\usage{
sc.cat.binning(data, target, n = 10, unique = 4, parallel = FALSE)
}
\arguments{
\item{data}{A data frame which contains target varible as well as predictor variables.}

\item{target}{Target variable name.}

\item{n}{Number of bootstrap iterations. Default 10 times.}

\item{unique}{The minimum number of unique values within predictor variables. Default 4.}

\item{parallel}{A logical scalar. Use parallel backend. Default FALSE.}
}
\value{
The output is a list of group plan which can be applied to the orginal data frame via
  the \code{\link{predict}} function.
  The user can also update the cut plan via the \code{\link{update}} function.
}
\description{
This is an easy categorical binning solution for credit scorecard build. It is designed to
group the optimal categories by utilizing the \href{https://cran.r-project.org/package=partykit}{Recursive Partitioning}
which is only applied on factor variables.
}
\examples{
\dontrun{
# Load library
library(easysc)

# Generate a grouping plan which maximize IV via 500 bootstrap resampling
group.plan <- sc.cat.binning(data = df, target = BAD, n = 500, parallel = TRUE)
# Update the grouping plan
update(group.plan, MARRIAGE = list(c('SINGLE', 'DIVORCE'),
                                   c('MARRIED', 'WIDOW'))
# Apply to the data frame
predict(group.plan, df, keepTarget = TRUE)
}
}
