% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMiscellaneous.R
\name{sc.Misc}
\alias{sc.Misc}
\alias{sc.score}
\alias{sc.point}
\alias{sc.corplot}
\alias{sc.performance}
\title{Miscellaneous tools for Credit Scoring}
\usage{
sc.score(data, model, pdo, score, odd)

sc.point(data, woe, model, pdo, score, odd)

sc.corplot(model, data, s = NULL)

sc.performance(model, X, y, s = NULL, plot = F)
}
\arguments{
\item{data}{A data frame which contains target varible as well as predictor variables.}

\item{model}{A model built.}

\item{pdo}{Point double odds.}

\item{score}{Score at which the desired odd is.}

\item{odd}{The desired odd.}

\item{woe}{A woe result which is calculated by \code{\link[klaR:woe]{woe}}.}

\item{s}{In case of using \code{\link[glmnet:cv.glmnet]{Lasso}} model. Lambda min or
lambda at 1se is chosen.}

\item{X}{A data frame which has only predictor variables.}

\item{y}{A vector of target variable.}

\item{plot}{A logical scalar. Plot ROC curve.}
}
\value{
\code{\link{sc.score}}. The output is a data frame with chosen variables in the model. All variables are in score point.
  There is also a final score.

\code{\link{sc.point}}. The output is a list with chosen variables in the model. Variables with their values and score point.

\code{\link{sc.corplot}}. The output plot is generated showing correlated variables.

\code{\link{sc.performance}}. The output GINI & KS is printed.
}
\description{
These are set of tools for easy credit scorecard build. The set includes \code{\link{sc.score}},
\code{\link{sc.point}} and \code{\link{sc.corplot}}.
}
\examples{
\dontrun{
# Load library
library(easysc)

# Generate final score data based on the model built with data.
scoredata <- sc.score(data = df, model = model.Lasso, pdo = 100, score = 800, odd = 5)

# Generate final scorecard based on the model built with data.
scorecard <- sc.point(data = df, woe = woe, model = model.Lasso, pdo = 100, score = 800, odd = 5)
}
}
